Fix debug logging to check once on first time logging is used, bugfix, and cleanup logging
