+++ src/dcap_provider.cpp
@@ -19,6 +19,7 @@
 #include <unordered_map>
 
 #include "sgx_ql_lib_common.h"
+#include "environment.h"
 
 #ifdef __LINUX__
 #include <arpa/inet.h>
@@ -26,18 +27,6 @@
 #include <intsafe.h>
 #include <winsock.h>
 #endif
-
-// NOTE:
-//  The environment variables below are mostly meant to be modified
-//  by the OE Jenkins environment to support CI/CD testing. Do not
-//  modify or override these values as they can cause regressions in
-//  caching service behavior.
-#define ENV_AZDCAP_BASE_URL "AZDCAP_BASE_CERT_URL"
-#define ENV_AZDCAP_CLIENT_ID "AZDCAP_CLIENT_ID"
-#define ENV_AZDCAP_COLLATERAL_VER "AZDCAP_COLLATERAL_VERSION"
-#define ENV_AZDCAP_DEBUG_LOG "AZDCAP_DEBUG_LOG_LEVEL"
-#define ENV_AZDCAP_DISABLE_ONDEMAND "AZDCAP_DISABLE_ONDEMAND"
-#define MAX_ENV_VAR_LENGTH 2000
 
 // External function names are dictated by Intel
 // ReSharper disable CppInconsistentNaming
@@ -90,55 +79,6 @@ enum class CollateralTypes
     PckCrl,
     PckRootCrl
 };
-
-static std::string get_env_variable(std::string env_variable)
-{
-    const char* env_value;
-#ifdef __LINUX__
-    env_value = getenv(env_variable.c_str());
-    if (env_value == NULL)
-    {
-        return std::string();
-    }
-#else
-    std::unique_ptr<char[]> env_temp =
-        std::make_unique<char[]>(MAX_ENV_VAR_LENGTH);
-    if (env_temp == nullptr)
-    {
-        log(SGX_QL_LOG_ERROR,
-            "Failed to allocate memory for environment varible for '%s'",
-            env_variable.c_str(),
-            MAX_ENV_VAR_LENGTH);
-    }
-    env_value = env_temp.get();
-    DWORD status = GetEnvironmentVariableA(
-        env_variable.c_str(), env_temp.get(), MAX_ENV_VAR_LENGTH);
-    if (status == 0)
-    {
-        log(SGX_QL_LOG_ERROR,
-            "Failed to retreive environment varible for '%s'",
-            env_variable.c_str(),
-            MAX_ENV_VAR_LENGTH);
-        return std::string();
-    }
-#endif
-    else
-    {
-        if ((strnlen(env_value, MAX_ENV_VAR_LENGTH) <= 0) ||
-            (strnlen(env_value, MAX_ENV_VAR_LENGTH) == MAX_ENV_VAR_LENGTH))
-        {
-            log(SGX_QL_LOG_ERROR,
-                "Value specified in environment variable %s is either empty or "
-                "expected max length '%d'.",
-                env_variable.c_str(),
-                MAX_ENV_VAR_LENGTH);
-            return std::string();
-        }
-
-        auto retval = std::string(env_value);
-        return retval;
-    }
-}
 
 static std::string get_collateral_version()
 {
@@ -871,7 +811,7 @@ static std::string build_eppid_json(const sgx_ql_pck_cert_id_t& pck_cert_id)
     {
         if (disable_ondemand == "1")
         {
-            log(SGX_QL_LOG_WARNING, "On demand registration disabled. No eppid being sent to cache");
+            log(SGX_QL_LOG_WARNING, "On demand registration disabled by environment variable. No eppid being sent to caching service");
             return "";
         }
     }
@@ -881,12 +821,12 @@ static std::string build_eppid_json(const sgx_ql_pck_cert_id_t& pck_cert_id)
 
     if (eppid.empty())
     {
-        log(SGX_QL_LOG_WARNING, "No eppid provided");
+        log(SGX_QL_LOG_WARNING, "No eppid provided - unable to send to caching service");
         return "";
     }
     else 
     {
-        log(SGX_QL_LOG_INFO, "Sending eppid provided");
+        log(SGX_QL_LOG_INFO, "Sending the provided eppid to caching service");
     }
 
     static const char json_prefix[] = "{\"eppid\":\"";
@@ -1376,14 +1316,6 @@ extern "C" void sgx_ql_free_revocation_info(
 extern "C" sgx_plat_error_t sgx_ql_set_logging_function(
     sgx_ql_logging_function_t logger)
 {
-    if (logger_callback == nullptr)
-    {
-        auto log_level = get_env_variable(ENV_AZDCAP_DEBUG_LOG);
-        if (!log_level.empty())
-        {
-            enable_debug_logging(log_level);
-        }
-    }
     logger_callback = logger;
     return SGX_PLAT_ERROR_OK;
 }
